combine skills and certificates into expertise and delete them.
