+++ .todo
@@ -11,6 +11,8 @@
 - rename the hobbies toggle properly
 - Update carousel images for cycling hobby
 - update the description for each carousel slide 
+
+- change icon for website after 
  
 backend hosting:
 - rails
